Split ring_buffer into a library.
This will make it easier to unit test.
