+++ src/transporter.cpp
@@ -35,7 +35,6 @@
 // but modified to use a ring buffer, fix a few bugs, and split the UART
 // implementation to a separate file.
 
-#include <algorithm>
 #include <array>
 #include <cerrno>
 #include <cmath>
@@ -49,7 +48,6 @@
 #include <fcntl.h>
 #include <poll.h>
 #include <termios.h>
-#include <unistd.h>
 
 #include "ros2_serial_example/transporter.hpp"
 
@@ -58,202 +56,6 @@ namespace ros2_to_serial_bridge
 
 namespace transport
 {
-
-namespace impl
-{
-
-RingBuffer::RingBuffer(size_t capacity)
-{
-    size = capacity + 1;
-    buf = new uint8_t[size];
-    // TODO(clalancette): check for nullptr
-    head = tail = buf;
-}
-
-RingBuffer::~RingBuffer()
-{
-    delete [] buf;
-}
-
-size_t RingBuffer::buffer_size() const
-{
-    return size;
-}
-
-size_t RingBuffer::capacity() const
-{
-    return size - 1;
-}
-
-uint8_t *RingBuffer::end() const
-{
-    return buf + size;
-}
-
-size_t RingBuffer::bytes_free() const
-{
-    if (head >= tail)
-    {
-        return capacity() - (head - tail);
-    }
-
-    return tail - head - 1;
-}
-
-size_t RingBuffer::bytes_used() const
-{
-    return capacity() - bytes_free();
-}
-
-bool RingBuffer::is_full() const
-{
-    return bytes_free() == 0;
-}
-
-bool RingBuffer::is_empty() const
-{
-    return bytes_free() == capacity();
-}
-
-uint8_t *RingBuffer::nextp(uint8_t *p)
-{
-    // Given a ring buffer rb and a pointer to a location within its
-    // contiguous buffer, return the a pointer to the next logical
-    // location in the ring buffer.
-    return buf + ((++p - buf) % buffer_size());
-}
-
-ssize_t RingBuffer::read(int fd)
-{
-    uint8_t *bufend = end();
-    size_t nfree = bytes_free();
-
-    size_t count = bufend - head;
-    ssize_t n = ::read(fd, head, count);
-    if (n > 0)
-    {
-        head += n;
-
-        // wrap?
-        if (head == bufend)
-        {
-            head = buf;
-        }
-
-        // fix up the tail pointer if an overflow occurred
-        if (static_cast<size_t>(n) > nfree)
-        {
-            tail = nextp(head);
-        }
-    }
-
-    return n;
-}
-
-void *RingBuffer::peek(void *dst, size_t count)
-{
-    if (count > bytes_used())
-    {
-        return nullptr;
-    }
-
-    uint8_t *u8dst = static_cast<uint8_t *>(dst);
-    uint8_t *bufend = end();
-    size_t nwritten = 0;
-    uint8_t *tmptail = tail;
-    while (nwritten != count)
-    {
-        size_t n = std::min(static_cast<size_t>(bufend - tmptail), count - nwritten);
-        ::memcpy(u8dst + nwritten, tmptail, n);
-        tmptail += n;
-        nwritten += n;
-
-        // wrap?
-        if (tmptail == bufend)
-        {
-            tmptail = buf;
-        }
-    }
-
-    return tmptail;
-}
-
-void *RingBuffer::memcpy_from(void *dst, size_t count)
-{
-    if (count > bytes_used())
-    {
-        return nullptr;
-    }
-
-    uint8_t *u8dst = static_cast<uint8_t *>(dst);
-    uint8_t *bufend = end();
-    size_t nwritten = 0;
-    while (nwritten != count)
-    {
-        size_t n = std::min(static_cast<size_t>(bufend - tail), count - nwritten);
-        ::memcpy(u8dst + nwritten, tail, n);
-        tail += n;
-        nwritten += n;
-
-        // wrap?
-        if (tail == bufend)
-        {
-          tail = buf;
-        }
-    }
-
-    return tail;
-}
-
-size_t RingBuffer::findseq(uint8_t *seq, size_t seqlen)
-{
-    size_t used = bytes_used();
-    if (used < seqlen)
-    {
-        // There aren't enough bytes in the ring for this sequence, so it
-        // can't possibly contain the entire sequence.
-        return used;
-    }
-
-    if (seqlen > capacity())
-    {
-        // The sequence to look for is larger than we can possibly hold;
-        // this can't work.
-        return used;
-    }
-
-    uint8_t *ringp = tail;
-    uint8_t *seqp = seq;
-    uint8_t *found = nullptr;
-    uint8_t *start = buf + ((tail - buf) % buffer_size());
-
-    while (ringp != head)
-    {
-        if (*ringp == *seqp)
-        {
-            if (found != nullptr && seqp == (seq + seqlen - 1))
-            {
-                // Found it!  Return the start of the sequence
-                return found - start;
-            }
-
-            if (found == nullptr)
-            {
-                found = ringp;
-            }
-            seqp++;
-        }
-        else
-        {
-            found = nullptr;
-        }
-        ringp = nextp(ringp);
-    }
-
-    return used;
-}
-
-}  // namespace impl
 
 /** CRC table for the CRC-16. The poly is 0x8005 (x^16 + x^15 + x^2 + 1) */
 uint16_t const crc16_table[256] = {
@@ -510,4 +312,5 @@ ssize_t Transporter::write(const topic_id_size_t topic_ID, char buffer[], size_t
 }
 
 }  // namespace transport
+
 }  // namespace ros2_to_serial_bridge
